Add files via upload
PROBLEM STATEMENT:
HELP International is an international humanitarian NGO that is committed to fighting poverty and providing the people of backward countries with basic amenities and relief during the time of disasters and natural calamities. It runs a lot of operational projects from time to time along with advocacy drives to raise awareness as well as for funding purposes.

The job is to categorise the countries using some socio-economic and health factors that determine the overall development of the country.