Add files via upload
You work for an e-commerce company ABC, and you’ve been asked to analyze and present your findings to the Sales and the Operations heads. The solution should include a summary of current state of the business, business and customer trends that you observe from the data and a proposal (if any) for areas where you think the company can improve.