Add files via upload
To identify patterns which indicate if a client has difficulty paying their installments which may be used for taking actions such as denying the loan, reducing the amount of loan, lending (to risky applicants) at a higher interest rate, etc.

The company wants to understand the driving factors (or driver variables) behind loan default, i.e. the variables which are strong indicators of default.