Add files via upload
The problem statement is to predict fraudulent credit card transactions with the help of machine learning models.
The dataset is taken from the Kaggle website and it has a total of 2,84,807 transactions, out of which 492 are fraudulent. Since the dataset is highly imbalanced, so it needs to be handled before model building.